Convert to src layout
